+++ coloraide/color/__init__.py
@@ -35,6 +35,8 @@ from ..spaces.din99o import Din99o
 from ..spaces.din99o_lch import Din99oLch
 from ..spaces.luv import Luv
 from ..spaces.lchuv import Lchuv
+from ..spaces.okhsl import Okhsl
+from ..spaces.okhsv import Okhsv
 from .distance import DeltaE
 from .distance.delta_e_76 import DE76
 from .distance.delta_e_94 import DE94
@@ -55,7 +57,8 @@ SUPPORTED_DE = (
 SUPPORTED_SPACES = (
     HSL, HWB, Lab, Lch, LabD65, LchD65, SRGB, SRGBLinear, HSV,
     DisplayP3, A98RGB, ProPhotoRGB, Rec2020, XYZ, XYZD65,
-    Oklab, Oklch, Jzazbz, JzCzhz, ICtCp, Din99o, Din99oLch, Luv, Lchuv
+    Oklab, Oklch, Jzazbz, JzCzhz, ICtCp, Din99o, Din99oLch, Luv, Lchuv,
+    Okhsl, Okhsv
 )
 
 SUPPORTED_FIT = (